api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-multipart==0.0.20
google-generativeai>=0.3.0
